// frontend/src/api/model/githubCom4H1RZooraInternalDomainLeadStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainLeadStatus = typeof GithubCom4H1RZooraInternalDomainLeadStatus[keyof typeof GithubCom4H1RZooraInternalDomainLeadStatus];


export const GithubCom4H1RZooraInternalDomainLeadStatus = {
  LeadStatusNew: 'new',
  LeadStatusContacted: 'contacted',
  LeadStatusConverted: 'converted',
  LeadStatusRejected: 'rejected',
} as const;